[MSGINA] Allow to pair FriendlyShutDownDlg with 'Server' CORE-17282
A ros specific addition that does not exist on Windows.

Default for Server is still the classic shutdown dlg,
no change on that.

But interested users can manually now create the registry setting
"EnforceFriendlyShutdown" REG_DWORD to "1"
within "SYSTEM\\CurrentControlSet\\Control\\Windows"
to allow overriding the Windows default behavior
without relying on changing the product version number, because
that would not be side-effect-free.

Addendum to 0.4.15-dev-650-g
2edcb58e6550fd6283cf3a7d332ee61849704f3d